Move PsbtSighashType to its own module
In preparation for adding PSBT v2, move the `PsbtSighashType` to its own
module in the crate root.
