+++ src/map/mod.rs
@@ -10,7 +10,7 @@ use crate::serialize::Serialize;
 
 #[rustfmt::skip]                // Keep pubic re-exports separate
 pub use self::{
-    input::{Input, PsbtSighashType},
+    input::Input,
     output::Output,
     global::Global,
 };
